Ktu-Notes-Telegram-Bot/requirements.txt

aiogram>=3.22.0
motor>=3.7.1
pymongo>=4.14.1
python-dotenv>=1.0.0
dnspython>=2.0.0
